+++ models/src/main/kotlin/io/provenance/api/models/p8e/tx/permission/DataAccessChangeType.kt
@@ -1,0 +1,6 @@
+package io.provenance.api.models.p8e.tx.permission
+
+enum class DataAccessChangeType {
+    ADD,
+    REMOVE
+}
