+++ models/src/main/kotlin/io/provenance/api/models/p8e/tx/permission/DataAccessUpdate.kt
@@ -1,0 +1,6 @@
+package io.provenance.api.models.p8e.tx.permission
+
+data class DataAccessUpdate(
+    val type: DataAccessChangeType,
+    val address: String,
+)
