+++ service/src/main/kotlin/io/provenance/api/domain/usecase/provenance/tx/permission/models/UpdateScopeDataAccessRequestWrapper.kt
@@ -1,6 +1,6 @@
 package io.provenance.api.domain.usecase.provenance.tx.permission.models
 
-import io.provenance.api.models.p8e.tx.UpdateScopeDataAccessRequest
+import io.provenance.api.models.p8e.tx.permission.UpdateScopeDataAccessRequest
 import java.util.UUID
 
 data class UpdateScopeDataAccessRequestWrapper(
